+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 --------------------------
 - Fix #85: Fix downloading of large user export data file
 - Enh #90: Use PHP CS Fixer
+- Fix #95: Fix Color Variables for Cookie Consent Box
 
 1.4.2 (September 13, 2024)
 --------------------------