+++ docs/CHANGELOG.md
@@ -1,8 +1,8 @@
 Changelog
 =========
 
-1.4.3 (Unreleased)
---------------------------
+1.4.3 (March 10, 2025)
+----------------------
 - Fix #85: Fix downloading of large user export data file
 - Enh #90: Use PHP CS Fixer
 - Fix #95: Fix Color Variables for Cookie Consent Box